Move charge reference data to its own file
We also update it to derive as much as possible from the licence data passed in. This helps ensure what we build is 'real'. If you seed the unregistered licence scenario, and add a charge version using abstraction data manually, what it creates matches what we seed using these `charge-version` and `charge-reference` data builders.
