+++ tests/support/data/charge-version.data.js
@@ -1,8 +1,7 @@
+import { regionCode } from '../default-values.js'
 import { generateUUID } from '../helpers/generate-uuid.js'
 
 export default function (billingAccountData, licenceData) {
-  const chargeElementId = generateUUID()
-  const chargeReferenceId = generateUUID()
   const chargeVersionId = generateUUID()
 
   const {
@@ -30,72 +29,8 @@ export default function (billingAccountData, licenceData) {
           schema: 'water',
           table: 'changeReasons',
           lookup: 'description',
-          value: 'Strategic review of charges (SRoC)',
+          value: 'New licence',
           select: 'changeReasonId'
-        }
-      }
-    ],
-    chargeReferences: [
-      {
-        id: chargeReferenceId,
-        chargeVersionId,
-        description: 'SROC Charge Reference 01',
-        source: 'tidal',
-        loss: 'high',
-        factorsOverridden: false,
-        chargeCategoryId: {
-          schema: 'public',
-          table: 'chargeCategories',
-          lookup: 'reference',
-          value: '4.3.1',
-          select: 'id'
-        },
-        adjustments: {
-          aggregate: null,
-          s126: null,
-          s127: true,
-          s130: false,
-          charge: null,
-          winter: false
-        },
-        waterModel: 'no model',
-        volume: 32,
-        eiucRegion: 'Southern',
-        section127Agreement: true
-      }
-    ],
-    chargeElements: [
-      {
-        id: chargeElementId,
-        chargeReferenceId,
-        abstractionPeriodStartDay: 1,
-        abstractionPeriodStartMonth: 4,
-        abstractionPeriodEndDay: 31,
-        abstractionPeriodEndMonth: 3,
-        authorisedAnnualQuantity: 32,
-        loss: 'high',
-        section127Agreement: true,
-        description: 'SROC Charge Purpose 01',
-        purposeId: {
-          schema: 'public',
-          table: 'purposes',
-          lookup: 'legacyId',
-          value: '400',
-          select: 'id'
-        },
-        purposePrimaryId: {
-          schema: 'water',
-          table: 'purposesPrimary',
-          lookup: 'legacyId',
-          value: 'A',
-          select: 'purposePrimaryId'
-        },
-        purposeSecondaryId: {
-          schema: 'water',
-          table: 'purposesSecondary',
-          lookup: 'legacyId',
-          value: 'AGR',
-          select: 'purposeSecondaryId'
         }
       }
     ]
